Fix: update JointController.py and ServoConnection.py

+++ ServoConnection.py
@@ -16,7 +16,7 @@ class ServoConnection:
     def SyncWriteServos(self,
                         servosIds: List[int],
                         servosSpeed:  List[int],
-                        goalsPos:  List[float]):
+                        goalsPos:  List[int]):
         
         scs_error = None
         scs_add_param_result = None
@@ -46,7 +46,7 @@ class ServoConnection:
     def SyncWriteServo(self,
                        servoId: int,
                        servoSpeed: int,
-                       goalPos: float):
+                       goalPos: int):
 
         scs_comm_result, scs_error = self.packetHandler.write2ByteTxRx(self.portHandler, servoId,
                                                                        self.ADDR_STS_GOAL_SPEED,
